Add (temporary) code to manually add species Sedum pallidum to the GRIIS checklist
This is a temporary solution as we source this script to get the GRIIS checklist data on the alien species portal and to construct the cubes for the alien species portal. As I will use these cubes for the project that is currently running with Wallonia, it would be good to have Sedum pallidum on the list. As soon as this species is added to the actual GRIIS checklist, I will remove this code again. I'm aware that a lot of data fields will be empty but having the species in there in this way will at least make sure that the occurrence data are downloaded for the cubes.